will display weather if given the api key
